+++ exof/src/main/java/team/balam/exof/Constant.java
@@ -9,8 +9,6 @@ public class Constant {
 	public static final String NETWORK_CHARSET = "UTF-8";
 	public static final String YES = "yes";
 	public static final String NO = "no";
-	public static final String FAIL = "fail";
-	public static final String SUCCESS = "success";
 	public static final String EMPTY_STRING = "";
 
 	public static final String ENV_DB = ".env.db";
