--- exof/src/main/java/team/balam/exof/Module.java
@@ -1,13 +1,0 @@
-package team.balam.exof;
-
-/**
- * Framework를 구성하는 요소
- * @author kwonsm
- *
- */
-public interface Module
-{
-	void start() throws Exception;
-	
-	void stop() throws Exception;
-}
