+++ exof/src/main/java/team/balam/exof/module/Module.java
@@ -1,0 +1,13 @@
+package team.balam.exof.module;
+
+/**
+ * Framework를 구성하는 요소
+ * @author kwonsm
+ *
+ */
+public interface Module
+{
+	void start() throws Exception;
+	
+	void stop() throws Exception;
+}
