+++ exof/src/main/java/team/balam/exof/module/was/JettyModule.java
@@ -9,7 +9,7 @@ import org.eclipse.jetty.util.ssl.SslContextFactory;
 import org.eclipse.jetty.webapp.WebAppContext;
 
 import team.balam.exof.Constant;
-import team.balam.exof.Module;
+import team.balam.exof.module.Module;
 import team.balam.exof.environment.EnvKey;
 import team.balam.exof.environment.SystemSetting;
 import team.balam.exof.module.listener.PortInfo;
